docs: add template for documentation (#42)
This PR adds a documentation template file to establish the structure for project documentation. The template includes a basic header with project branding and a placeholder for README content that is automatically applied in the [aweXpect Build.Pages](https://github.com/aweXpect/aweXpect/blob/main/Pipeline/Build.Pages.cs) step.